metodo process de enviar dados de pagamento para o raspay finalizando o envio de dados de pagamento , iniciando envio de email

+++ src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDTO.java
@@ -14,6 +14,6 @@ import java.math.BigDecimal;
 public class OrderDTO {
     private String id;
     private String costumerId;
-    private Long discount;
+    private BigDecimal discount;
     private String productAcronym;
 }
